# wacp-console/openapi.yaml
info:
  description: REST API for the WACP Console — coordination workbench for the WACP ecosystem.
  license:
    name: Apache-2.0
  title: WACP Console API
  version: 0.1.0
openapi: 3.1.0
paths:
...
  /api/sessions/{id}/ws:
    get:
      responses:
        '101':
          description: Response
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: WebSocket upgrade
      tags:
      - sessions
...
  /api/users:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: List users
      tags:
      - users
    post:
      responses:
        '201':
          description: Created
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Create user
      tags:
      - users
  /api/users/{id}:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get user
      tags:
      - users
    patch:
      responses:
        '204':
          description: No content
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Update user
      tags:
      - users
  /api/users/{id}/reset-password:
    post:
      responses:
        '204':
          description: No content
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Reset user password
      tags:
      - users
  /api/users/{id}/unlock:
    post:
      responses:
        '204':
          description: No content
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Unlock user
      tags:
      - users
...
  /api/verticals/{id}/checkpoint-types:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get vertical checkpoint types
      tags:
      - verticals
  /api/verticals/{id}/context-schema:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get context schema
      tags:
      - verticals
  /api/verticals/{id}/quality-criteria:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get quality criteria
      tags:
      - verticals
  /api/verticals/{id}/task-types:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: List task types
      tags:
      - verticals
  /api/verticals/{id}/tool-policies:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get tool policies
      tags:
      - verticals
  /api/verticals/{id}/workflows:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: List workflows
      tags:
      - verticals
  /api/verticals/{id}/workflows/{wf_id}:
    get:
      responses:
        '200':
          description: Success
        '401':
          description: Unauthenticated
        '403':
          description: Forbidden
      summary: Get workflow
      tags:
      - verticals
...